third_party/blink/web_tests/fast/dom/collection-length-should-not-be-overridden-expected.txt

This test ensures the builtin length property of HTMLCollection's subclasses is not overridden by a named getter of the same name.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS documentAllLength is 18
PASS documentAll5 is documentAllSome
PASS documentAll6.name is 'length'
PASS form.length is form.elements[1]
PASS form.elements.length is 2
PASS form.elements[0] is form.elements['some']
PASS form.elements[1].name is 'length'
PASS select.options.length is 2
PASS select.options[0].getAttribute('name') is 'length'
PASS select.options[1] is select.options['some']
PASS table.rows.length is 2
PASS table.rows[0] is table.rows['another']
PASS table.rows[1].id is 'length'
PASS successfullyParsed is true

TEST COMPLETE
